Try to remove effects.js middleware
