+++ packages/block-editor/src/store/index.js
@@ -7,7 +7,6 @@ import { createReduxStore, registerStore } from '@wordpress/data';
  * Internal dependencies
  */
 import reducer from './reducer';
-import applyMiddlewares from './middlewares';
 import * as selectors from './selectors';
 import * as actions from './actions';
 import controls from './controls';
